% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relicoef.R
\name{relicoef}
\alias{relicoef}
\title{Raykov's factor reliability coefficient}
\usage{
relicoef(mod)
}
\description{
This function computes Raykov's reliability coefficient (RRC) for factors from
confirmatory factor analyses, a measure which is commonly seen as a more
accurate one than that of Cronbach's alpha which is computed based on the
assumption of tau-equivalent measures. It computes reliability coefficients
for factors with and without correlated errors.
}
\examples{
model.01 <- '
           Col =~ respected + secure
           Ind =~ accomplish + self_fulfil + self_respect
           '
est.model.01 <- cfa(model.01, data=values)
relicoef(est.model.01)

}
